# docs/swagger.yaml
basePath: /
definitions:
  models.CwmpConfig:
    properties:
      content:
        description: script content
        type: string
      created_at:
        type: string
      id:
        description: 主键 ID
        type: string
      level:
        description: script level  normal｜major
        type: string
      name:
        description: Name
        type: string
      oid:
        type: string
      oui:
        type: string
      product_class:
        type: string
      software_version:
        type: string
      target_filename:
        type: string
      task_tags:
        description: task label
        type: string
      timeout:
        description: Execution Timeout Seconds
        type: integer
      updated_at:
        type: string
    type: object
  models.CwmpConfigSession:
    properties:
      config_id:
        description: Script ID
        example: ""
        type: string
      content:
        description: script content
        type: string
      cpe_id:
        description: CPE ID
        example: "0"
        type: string
      created_at:
        type: string
      exec_status:
        description: execution state  success | failure
        type: string
      exec_time:
        description: execution time
        type: string
      id:
        description: primary key ID
        example: "0"
        type: string
      last_error:
        description: last execution error
        type: string
      level:
        description: script level  normal｜major
        type: string
      name:
        description: Name
        type: string
      oui:
        type: string
      product_class:
        type: string
      resp_time:
        description: Response time
        type: string
      session:
        description: Session ID
        type: string
      software_version:
        type: string
      task_tags:
        description: task label
        type: string
      timeout:
        description: execution timeout second
        type: integer
      updated_at:
        type: string
    type: object
  models.CwmpPresetTask:
    properties:
      batch:
        type: string
      content:
        type: string
      created_at:
        type: string
      event:
        type: string
      exec_time:
        description: 执行时间
        type: string
      id:
        description: 主键 ID
        example: "0"
        type: string
      name:
        type: string
      oid:
        type: string
      onfail:
        type: string
      preset_id:
        example: "0"
        type: string
      request:
        type: string
      resp_time:
        description: 响应时间
        type: string
      response:
        type: string
      session:
        type: string
      sn:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  models.NetCpe:
    properties:
      arch_name:
        description: 设备架构
        type: string
      cpe_type:
        description: 设备类型
        type: string
      cpu_usage:
        description: CPE 百分比
        type: integer
      created_at:
        type: string
      cwmp_last_inform:
        description: CWMP 最后检测时间
        type: string
      cwmp_status:
        description: cwmp 状态
        type: string
      cwmp_url:
        type: string
      factoryreset_id:
        type: string
      hardware_version:
        description: 设备版本
        type: string
      id:
        description: 主键 ID
        example: "0"
        type: string
      manufacturer:
        description: 设备制造商
        type: string
      memory_free:
        description: 内存可用
        type: integer
      memory_total:
        description: 内存总量
        type: integer
      model:
        description: 设备型号
        type: string
      name:
        description: 设备名称
        type: string
      node_id:
        description: 节点ID
        example: "0"
        type: string
      oui:
        description: 设备OUI
        type: string
      product_class:
        description: 设备类型
        type: string
      remark:
        description: 备注
        type: string
      sn:
        description: 设备序列号
        type: string
      software_version:
        description: 设备固件版本
        type: string
      status:
        description: 设备状态
        type: string
      system_name:
        description: 设备系统名称
        type: string
      tags:
        description: 标签
        type: string
      task_tags:
        description: 任务标签
        type: string
      updated_at:
        type: string
      uptime:
        description: UpTime
        type: integer
      vendor_code:
        description: 设备厂商代码
        type: string
    type: object
host: 127.0.0.1:1816
info:
  contact:
    email: jamiesun.net@gmail.com
    name: Toughradius API Support
    url: https://github.com/talkincode/toughradius
  description: This is Toughradius API
  license:
    name: GPL
    url: https://github.com/talkincode/toughradius
  termsOfService: https://github.com/talkincode/toughradius
  title: Toughradius API
  version: "1.0"
paths:
  /api/cpe/query:
    get:
      consumes:
      - application/json
      description: Query cpe list
      parameters:
      - description: node_id
        in: query
        name: node_id
        type: string
      - description: customer_id
        in: query
        name: customer_id
        type: string
      - description: keyword
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.NetCpe'
            type: array
      security:
      - BearerAuth: []
      summary: Query CPE list
      tags:
      - CPE
  /api/cwmp/config/query:
    get:
      consumes:
      - application/json
      description: Query cwmp config list
      parameters:
      - description: keyword
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CwmpConfig'
            type: array
      security:
      - BearerAuth: []
      summary: Query cwmp config list
      tags:
      - TR069
  /api/cwmp/config/session/query:
    get:
      consumes:
      - application/json
      description: Query cwmp config sessions
      parameters:
      - description: cpe_id
        in: query
        name: cpe_id
        type: string
      - description: keyword
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CwmpConfigSession'
            type: array
      security:
      - BearerAuth: []
      summary: Query cwmp config sessions
      tags:
      - TR069
  /api/cwmp/preset/task/query:
    get:
      consumes:
      - application/json
      description: Query cwmp preset task
      parameters:
      - description: cpe_id
        in: query
        name: cpe_id
        type: string
      - description: keyword
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CwmpPresetTask'
            type: array
      security:
      - BearerAuth: []
      summary: Query cwmp preset task
      tags:
      - TR069
securityDefinitions:
  BearerAuth:
    description: Accesskey based security scheme to secure api
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"